Rename package from tssim to whorl
Comprehensive rename of the tool to better reflect its purpose as a
"whorl finder" in code - identifying clusters of similar code structures
branching from common points.

Changes include:
- Package name: tssim -> whorl
- CLI command: tssim -> whorl
- Directory structure: tssim/ -> whorl/
- All imports and references updated across Python files
- Documentation updated (README, .claude.md, ADRs)
- Configuration files updated (pyproject.toml, scripts)
- Test files and benchmark framework updated
- SARIF output metadata updated to reflect new tool name

All tests, linting, and type checking pass after the rename.
