+++ .claude.md
@@ -4,7 +4,7 @@ This file provides guidance to Claude Code (claude.ai/code) when working with co
 
 ## Project Overview
 
-**tssim** is a code similarity detection tool that identifies structurally similar code blocks across and within programming languages. Unlike traditional duplicate detection tools that only find textual matches, tssim uses Tree-sitter ASTs to detect semantic and structural similarities even when identifiers, literals, or function names differ.
+**whorl** is a code similarity detection tool that identifies structurally similar code blocks across and within programming languages. Unlike traditional duplicate detection tools that only find textual matches, whorl uses Tree-sitter ASTs to detect semantic and structural similarities even when identifiers, literals, or function names differ.
 
 ### Architecture
 
@@ -73,8 +73,8 @@ Writing guidance:
 ## Project Structure
 
 ```
-tssim/          # Main package source code
-tests/          # Test files (mirror tssim/ structure)
+whorl/          # Main package source code
+tests/          # Test files (mirror whorl/ structure)
 docs/adr/       # Architecture Decision Records
 ```
