+++ .github/scripts/check_radon.sh
@@ -6,7 +6,7 @@ set -u
 set -e
 
 echo "Radon Cyclomatic Complexity below grade A:"
-PACKAGE="tssim"
+PACKAGE="whorl"
 CC_COMMAND="uv run radon cc -n B ${PACKAGE}"
 CC=$($CC_COMMAND)
