+++ .nvimrc
@@ -5,7 +5,7 @@
 let g:gutentags_ctags_exclude += ['*/.venv/*', '*/benchmark-tests/*']
 let g:projectionist_heuristics = {
       \ 'pyproject.toml': {
-      \   'tssim/*.py': {
+      \   'whorl/*.py': {
       \     'type': 'function',
       \     'alternate': [
       \       'tests/{dirname}test_{basename}.py',
@@ -15,8 +15,8 @@ let g:projectionist_heuristics = {
       \   'tests/**/test_*.py': {
       \     'type': 'test',
       \     'alternate': [
-      \       'tssim/{dirname}{basename}.py',
-      \       'tssim/{dirname}/{basename}.py',
+      \       'whorl/{dirname}{basename}.py',
+      \       'whorl/{dirname}/{basename}.py',
       \     ]
       \   },
       \ },
