--- tssim/models/__init__.py
@@ -1,5 +1,0 @@
-"""Domain models for tssim."""
-
-from tssim.models.ast import ParsedFile, ParseResult
-
-__all__ = ["ParsedFile", "ParseResult"]
