--- tssim/pipeline/__init__.py
@@ -1,5 +1,0 @@
-"""Pipeline modules for tssim processing stages."""
-
-from tssim.pipeline.parse import parse_path
-
-__all__ = ["parse_path"]
