+++ whorl/formatters/__init__.py
@@ -1,0 +1,6 @@
+"""Output formatters for whorl results."""
+
+from whorl.formatters.sarif import format_as_sarif
+from whorl.formatters.json import format_as_json
+
+__all__ = ["format_as_sarif", "format_as_json"]
