+++ whorl/pipeline/__init__.py
@@ -1,0 +1,5 @@
+"""Pipeline modules for whorl processing stages."""
+
+from whorl.pipeline.parse import parse_path
+
+__all__ = ["parse_path"]
