+++ whorl/pipeline/languages/css.py
@@ -1,6 +1,6 @@
 """CSS language configuration."""
 
-from tssim.pipeline.rules.models import Rule, RuleAction
+from whorl.pipeline.rules.models import Rule, RuleAction
 
 from .base import LanguageConfig, RegionExtractionRule
