+++ whorl/pipeline/languages/javascript.py
@@ -1,6 +1,6 @@
 """JavaScript language configuration."""
 
-from tssim.pipeline.rules.models import Rule, RuleAction
+from whorl.pipeline.rules.models import Rule, RuleAction
 
 from .base import LanguageConfig, RegionExtractionRule
