+++ whorl/terminal_detect.py
